Move HTTP API classes into a common package tree
Resolves current situation where HTTP APIs are distributed throughout the source tree.

- `*Resource.java` => `api` package
- `*Info.java` => `api.types` package
