+++ sdk/scheduler/src/test/java/com/mesosphere/sdk/api/TaskResourceTest.java
@@ -1,7 +1,9 @@
-package com.mesosphere.sdk.scheduler.api;
+package com.mesosphere.sdk.api;
 
 import com.googlecode.protobuf.format.JsonFormat;
 import org.apache.mesos.Protos.*;
+
+import com.mesosphere.sdk.api.TaskResource;
 import com.mesosphere.sdk.offer.CommonTaskUtils;
 import com.mesosphere.sdk.offer.ResourceUtils;
 import com.mesosphere.sdk.scheduler.TaskKiller;
