+++ sdk/scheduler/src/test/java/com/mesosphere/sdk/api/types/StringPropertyDeserializerTest.java
@@ -1,5 +1,6 @@
-package com.mesosphere.sdk.state.api;
+package com.mesosphere.sdk.api.types;
 
+import com.mesosphere.sdk.api.types.StringPropertyDeserializer;
 import com.mesosphere.sdk.config.SerializationUtils;
 import org.junit.Before;
 import org.junit.Test;
